+++ src/main/java/com/spartabugkiller/ecommercebackofficeproject/customer/dto/request/CustomerRequest.java
@@ -4,8 +4,10 @@ import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
+@Setter
 @NoArgsConstructor
 public class CustomerRequest {
     @NotBlank(message = "이름은 필수입니다.")
@@ -15,6 +17,6 @@ public class CustomerRequest {
     @Email(message = "올바른 이메일 형식이 아닙니다.")
     private String email;
 
-    @NotBlank(message = "비밀번호는 필수입니다.")
-    private String password;
+    @NotBlank(message = "휴대폰번호는 필수입니다.")
+    private String phoneNumber;
 }
